feat: handle emotion for video and follow

+++ src/services/video.js
@@ -1,6 +1,8 @@
 const Video = require("../models/Video");
 const Like = require("../models/Like");
 const Love = require("../models/Love");
+const User = require("../models/User");
+const AppError = require("../utils/error");
 
 class VideoService {
   static addVideo = async ({ url, size, duration, userId, content }) => {
@@ -31,53 +33,44 @@ class VideoService {
 
   static likeVideo = async ({ videoId, userId }) => {
     const video = await Video.findById(videoId);
+    if (!video) throw new AppError("Infomation of video not valid");
+
     const findLike = await Like.findOne({
       user_id: userId,
       video_id: videoId,
     });
     if (findLike) {
-      video.like = video.like - 1;
-      await findLike.deleteOne();
+      return await findLike.deleteOne();
     } else {
-      video.like = video.like + 1;
-      await Like.create({
+      return await Like.create({
         user_id: userId,
         video_id: videoId,
       });
     }
-    await video.save();
   };
 
   static loveVideo = async ({ videoId, userId }) => {
     const video = await Video.findById(videoId);
+    if (!video) throw new AppError("Infomation of video not valid");
+
     const findLove = await Love.findOne({
       user_id: userId,
       video_id: videoId,
     });
     if (findLove) {
-      video.love = video.love - 1;
-      await findLove.deleteOne();
+      return await findLove.deleteOne();
     } else {
-      video.love = video.love + 1;
-      await Love.create({
+      return await Love.create({
         user_id: userId,
         video_id: videoId,
       });
     }
-    await video.save();
   };
 
   static getFavorite = async ({ userId }) => {
-    const likesRes = await Like.find({ user_id: userId })
-      .select("video_id -_id")
-      .lean();
-    const lovesRes = await Love.find({ user_id: userId })
-      .select("video_id -_id")
-      .lean();
+    const likes = await Like.distinct("video_id", { user_id: userId });
+    const loves = await Love.distinct("video_id", { user_id: userId });
 
-    const likes = likesRes.map((like) => like.video_id);
-    const loves = lovesRes.map((love) => love.video_id);
-    // console.log(likes, loves)
     return {
       likes,
       loves,
